% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_plans.R
\name{build_methods_plan}
\alias{build_methods_plan}
\title{Make a drake plan with forecasting methods to be compared}
\usage{
build_methods_plan()
}
\value{
a drake plan (i.e. a tibble) with the forecasting methods.
}
\description{
We list out the different forecasting methods (and the parameter
settings we wish to try) in the form of a drake plan. Note the use of
\code{\link[MATSS]{analysis_wrapper}} that takes a function that operates on
a single time series, and wraps it to run on each time series within a
dataset.
}
