% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pe_code.R
\name{check_time_series}
\alias{check_time_series}
\title{Check time series for potentially problematic features}
\usage{
check_time_series(x)
}
\arguments{
\item{x}{A time series of real numbers, represented as a numeric or integer
vector.}
}
\value{
nothing
}
\description{
Some of the analysis methods have issues handling NA values, duplicated
values or consecutive segments of identical values. This function spits out
some warnings on those.
}
\examples{
check_time_series(Nile)

}
