% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-subsetting-time-series.R
\name{compute_subset_range}
\alias{compute_subset_range}
\title{Compute the indices corresponding to a defined subset}
\usage{
compute_subset_range(bounds, n)
}
\arguments{
\item{bounds}{the vector with lower and upper bound, expressed as numeric
fractions of the time series in \link{0, 1}}

\item{n}{the totla length of the vector to subset}
}
\value{
the indices corresponding to the subset defined by \code{bounds} on a
vector of length n
}
\description{
Compute the indices corresponding to a defined subset
}
\examples{
compute_subset_range(c(0, 0.5), 10) # 1:5
compute_subset_range(c(0.5, 1), 10) # 6:10

}
