% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pe_code.R
\name{embedd}
\alias{embedd}
\title{Embed a time series}
\usage{
embedd(x, m, d = 1, as.matrix = TRUE, as.embed = TRUE)
}
\arguments{
\item{x}{the time series, observed at regular intervals.}

\item{m}{the number of dimensions to embed x into.}

\item{d}{the time delay}

\item{as.matrix:}{logical; whether to convert a 1-dimensional embedding to a
column matrix}

\item{as.embed:}{logical; should we return the embedded time series in the
order that embed() would?}
}
\value{
matrix of the embedded time series
}
\description{
Embed a time series
}
