% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting-ets.R
\name{ets_ts}
\alias{ets_ts}
\alias{ets_one_step}
\title{Exponentially smoothed time series model}
\usage{
ets_ts(timeseries, num_ahead = 5, level = 95, frequency = 1)

ets_one_step(timeseries, level = 95)
}
\arguments{
\item{timeseries}{the time series to forecast}

\item{num_ahead}{the number of points at the end of the time series to
forecast}

\item{level}{Confidence level for prediction intervals.}

\item{frequency}{the number of observations per unit of time.}
}
\value{
a data.frame of the mean forecasts, the observed values, and the
lower and upper CI levels (if an error occurs, then just NA values)
}
\description{
Fit a time series model using \code{\link[forecast]{ets}} and
make forecasts. The frequency of the data is set a priori, as opposed to
estimating the parameter from the data.

\code{ets_one_step} uses \code{\link[forecast]{ets}} to fit an
exponential-smoothing time series model and make a single one-step
forecast.
}
