% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-expectations.R
\name{expect_forecasts}
\alias{expect_forecasts}
\title{Check if object is in valid forecasts format}
\usage{
expect_forecasts(forecasts, columns = c("observed", "predicted"))
}
\arguments{
\item{forecasts}{the object of forecasts to be checked}

\item{columns}{the expecte columns for \code{forecasts} (defaults to
\code{columns = c("observed", "predicted")})}
}
\description{
Check if object is in valid forecasts format
}
