% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-expectations.R
\name{expect_NA_warnings}
\alias{expect_NA_warnings}
\title{Check if warnings are expected for a too-short time series}
\usage{
expect_NA_warnings(w)
}
\arguments{
\item{w}{the captured warnings}
}
\description{
Check if warnings are expected for a too-short time series
}
