% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting-utils.R
\name{forecast_one_step_static}
\alias{forecast_one_step_static}
\title{A wrapper function for one-step forecasting with a static model (no refitting)}
\usage{
forecast_one_step_static(fun, timeseries, num_ahead = 5, ...)
}
\arguments{
\item{fun}{a function for doing the forecasting. It should have arguments:
\describe{
\item{training}{the data for training the model}
\item{observed}{the observed values to be forecasted}
\item{...}{any optional arguments}
} and should return a data.frame with at least the observed and predicted
values (other columns are optional, and may be specific to the forecasting
method)}

\item{timeseries}{the time series to forecast}

\item{num_ahead}{the number of points at the end of the time series to
forecast}

\item{...}{arguments to pass to \code{fun}}
}
\value{
a data.frame of the observed and predicted values, with any other
remaining values returned by \code{fun}. If and error occurs, then NA values for
the observed and predicted
}
\description{
This function faciliates the forecasting approach in Ward et al.
2014. It has two operating modes. If \code{num_ahead} is set, then it forecasts
the last \code{num_ahead} points of the time series:
\enumerate{
\item fit the model to all but the last \code{num_ahead} time points
\item make 1-step ahead forecasts for the \code{num_ahead} time points
}
}
