% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting-wrapper.R
\name{forecast_wrapper}
\alias{forecast_wrapper}
\title{Apply a forecasting function to all time series in a dataset}
\usage{
forecast_wrapper(dataset, fun, ...)
}
\arguments{
\item{dataset}{the dataset object (following the \code{MATSS} data format)}

\item{fun}{the forecasting function}

\item{...}{extra params to be passed to the forecasting function}
}
\value{
a tibble with two list-columns:
\tabular{ll}{
\code{results} \tab a combined results data.frame\cr
\code{metadata} \tab the metadata component of the original dataset\cr
}
The results have all the columns returned from the forecasting function,
with additional columns for the variable that was forecast (pulled from
the original dataset), and a column for the forecast method
}
\description{
This wrapper applies a forecasting function to all the time
series in a dataset, and then coerces the results and metadata into a
combined object.
}
\examples{
\dontrun{
  salmon <- readd(data_.salmon., cache = cache)
  xx <- forecast_wrapper(salmon, autoarima_ts)
}

}
