% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting-gam.R
\name{gam_ts}
\alias{gam_ts}
\title{Make forecasts using a Generalized Additive Model}
\usage{
gam_ts(ts, num_ahead = 5, level = 95)
}
\arguments{
\item{level}{the CI level to include}
}
\value{
a data.frame of the mean forecasts, the observed values, and the
lower and upper CI levels (if an error occurs, then just NA values)
}
\description{
Fit a Generalized Additive Model (GAM) with spline over
time/years and use it to make forecasts
}
