% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_taxonomy.R
\name{get_taxonomy}
\alias{get_taxonomy}
\title{Get taxonomy data for all organisms}
\usage{
get_taxonomy()
}
\value{
a dataframe with all taxonomic ranks found for a particular species
}
\description{
Imports taxonomic information for all organisms that are forecasted
relying on functions of the \code{taxize} package
}
