% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-processing.R
\name{get_ward_data}
\alias{get_ward_data}
\title{Read in a specific database from Ward et al. 2014}
\usage{
get_ward_data(database_name, ward_RDS_file)
}
\arguments{
\item{database_name}{the name of the database}

\item{ward_RDS_file}{the location of the RDS file with the combined datasets
that is produced as a result of running \code{\link{reshape_ward_data}}}
}
\value{
the dataset object
}
\description{
Read in a specific database from Ward et al. 2014
}
