% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-forecasting.R
\name{hindcast}
\alias{hindcast}
\title{Hindcasting with one-step forecasts}
\usage{
hindcast(fun, timeseries, hindcast_method = c("pred_frac", "last_n",
  "pred_start"), pred_frac = 1/3, last_n = 5, pred_start = 31, ...)
}
\arguments{
\item{fun}{a function for doing the forecasting. It should have arguments:
\describe{
\item{training}{the data for training the model}
\item{observed}{optional observation of the next point}
\item{...}{any optional arguments}
} and should return a data.frame with at least the predicted forecast (in
a column named \code{predicted}). Other columns are optional, and may be
specific to the forecasting method.}

\item{timeseries}{the time series to forecast}

\item{hindcast_method}{the method for determining # of hindcasts}

\item{pred_frac}{the fraction of points at the end of the time series to
forecast}

\item{last_n}{the number of points at the end of the time series to
forecast}

\item{pred_start}{the index of the point of the time series at which to
begin forecasts}

\item{...}{arguments to pass to \code{fun}}
}
\value{
a data.frame of the observed and predicted values, with any other
remaining values returned by \code{fun}. If any error occurs, then NA values for
the observed and predicted
}
\description{
This function performs hindcasting over a specified segment of
the time series. The hindcasting is intended to simulate the use of a
forecasting method in practice. For a given time series of length \code{m},
suppose that the last \code{n} points are used for forecast evaluation, then
the procedure is:
\enumerate{
\item fit the model to the first \code{m - n} time points
\item make a 1-step ahead forecast
\item add the next observation to the training data, refit the model,
and make another forecast
\item repeat for the \code{n} forecasts
}

The value of \code{n} is determined in one of several ways, depending on the
\code{hindcast_selection} argument:
\describe{
\item{\code{last_n}}{make hindcasts for the last \code{last_n} points of the time
series}
\item{\code{pred_frac}}{make hindcasts for the last \code{pred_frac} fraction of
the time series}
\item{\code{pred_start}}{make hindcasts starting with the \code{pred_start} point
of the time series}
}
}
