% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting-lm.R
\name{lm_ts}
\alias{lm_ts}
\title{Make forecasts using a Linear Model}
\usage{
lm_ts(timeseries, num_ahead = 5, level = 95)
}
\arguments{
\item{timeseries}{the time series to forecast}

\item{num_ahead}{the number of points at the end of the time series to
forecast}

\item{level}{Confidence level for prediction intervals.}
}
\value{
a data.frame of the mean forecasts, the observed values, and the
lower and upper CI levels (if an error occurs, then just NA values)
}
\description{
Fit a Linear Regression for forecasting the time series as a
linear function over time/years
}
