% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting-locreg.R
\name{locreg_ts}
\alias{locreg_ts}
\title{Make forecasts using locally weighted regression}
\usage{
locreg_ts(timeseries, num_ahead = 5, level = 95, nn = seq(from = 0.2,
  to = 3, by = 0.05), deg = 1:3)
}
\arguments{
\item{timeseries}{the time series to forecast}

\item{num_ahead}{the number of points at the end of the time series to
forecast}

\item{level}{Confidence level for prediction intervals.}
}
\value{
a data.frame of the mean forecasts, the observed values, and the
lower and upper CI levels (if an error occurs, then just NA values)
}
\description{
\code{locreg_ts} fits a locally weighted regression. It first
searches over a parameter space of nearest neighbors bandwidth, and
polynomial degree, and then uses the best solution
}
