% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting-utils.R
\name{make_forecasts}
\alias{make_forecasts}
\title{A wrapper function for forecasting methods}
\usage{
make_forecasts(fun, timeseries, num_ahead = 5, ...)
}
\arguments{
\item{fun}{a function for doing the forecasting. It should have arguments:
\describe{
\item{training}{the data for training the model}
\item{observed}{the observed values to be forecasted}
\item{...}{any optional arguments}
} and should return a data.frame with at least the observed and predicted
values (other columns are optional, and may be specific to the forecasting
method)}

\item{timeseries}{the time series to forecast}

\item{num_ahead}{the number of points at the end of the time series to
forecast}

\item{...}{arguments to pass to \code{fun}}
}
\value{
a data.frame of the observed and predicted values, with any other
remaining values returned by \code{fun}. If and error occurs, then NA values for
the observed and predicted
}
\description{
This function faciliates the forecasting approach in Ward et al.
2014:
\enumerate{
\item fit the model to all but the last \code{num_ahead} time points
\item make 1-step ahead forecasts, and feed each forecast in as the
observed, until \code{num_ahead} total forecasts are made
}
}
