% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting-marss.R
\name{marss_ts}
\alias{marss_ts}
\title{Make forecasts using a state space model}
\usage{
marss_ts(timeseries, num_ahead = 5, level = 95, drift = TRUE,
  silent = TRUE)
}
\arguments{
\item{timeseries}{the time series to forecast}

\item{num_ahead}{the number of points at the end of the time series to
forecast}

\item{level}{Confidence level for prediction intervals.}

\item{drift}{whether the state-space model should include drift or not}

\item{silent}{ TRUE/FALSE Suppresses printing of full error messages, warnings, progress bars and convergence information. Setting silent=2 will produce more verbose error messages and progress information.}
}
\value{
a data.frame of the mean forecasts, the observed values, and the
lower and upper CI levels (if an error occurs, then just NA values)
}
\description{
Fit a state-space model using \code{\link[MARSS]{MARSS}} and
make forecasts.
}
