% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting-naive.R
\name{naive_one_step}
\alias{naive_one_step}
\title{Naive one-step ahead forecast}
\usage{
naive_one_step(timeseries, ...)
}
\arguments{
\item{timeseries}{the time series to forecast}

\item{...}{arguments to pass to \code{fun}}
}
\value{
a data.frame of the forecasts and the observed values
}
\description{
This model forecasts the next step ahead as the current value,
and is a baseline model for comparison against.
}
