% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting-npreg.R
\name{npreg_ts}
\alias{npreg_ts}
\title{Make forecasts using nonparametric kernel regression}
\usage{
npreg_ts(timeseries, num_ahead = 5, level = 95, regtype = "ll",
  bwmethod = "cv.aic", gradients = TRUE)
}
\arguments{
\item{timeseries}{the time series to forecast}

\item{num_ahead}{the number of points at the end of the time series to
forecast}

\item{level}{Confidence level for prediction intervals.}

\item{regtype}{
    a character string specifying which type of kernel regression
    estimator to use. \code{lc} specifies a local-constant estimator
    (Nadaraya-Watson) and \code{ll} specifies a local-linear
    estimator. Defaults to \code{lc}.
  }

\item{bwmethod}{
    which method to use to select bandwidths. \code{cv.aic} specifies
    expected Kullback-Leibler cross-validation (Hurvich, Simonoff, and
    Tsai (1998)), and \code{cv.ls} specifies least-squares
    cross-validation. Defaults to \code{cv.ls}.
  }

\item{gradients}{
    a logical value indicating that you want gradients computed and
    returned in the resulting \code{npregression} object. Defaults to
    \code{FALSE}.
  }
}
\value{
a data.frame of the mean forecasts, the observed values, and the
lower and upper CI levels (if an error occurs, then just NA values)
}
\description{
\code{npreg_ts} fits a timeseries model using \code{\link[np]{npreg}}
}
