% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-processing.R
\name{plan_ward_data}
\alias{plan_ward_data}
\title{Make a drake plan with all the datasets in Ward et al. 2014}
\usage{
plan_ward_data(ward_RDS_file = here::here("analysis", "data",
  "ward_fish_data.RDS"))
}
\arguments{
\item{ward_RDS_file}{the location of the RDS file with the combined datasets
that is produced as a result of running \code{\link{reshape_ward_data}}}
}
\value{
a drake plan (i.e. a tibble) specifying the targets and commands
for gathering datasets
}
\description{
Make a drake plan with all the datasets in Ward et al. 2014
}
