% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting-random-walk.R
\name{randomwalk_ts}
\alias{randomwalk_ts}
\alias{randomwalk_one_step}
\title{Forecast using a random walk model}
\usage{
randomwalk_ts(timeseries, num_ahead = 5, level = 95, drift = FALSE)

randomwalk_one_step(timeseries, level = 95, drift = FALSE)
}
\arguments{
\item{timeseries}{the time series to forecast}

\item{num_ahead}{the number of points at the end of the time series to
forecast}

\item{level}{Confidence levels for prediction intervals.}

\item{drift}{Logical flag. If TRUE, fits a random walk with drift model.}
}
\description{
\code{randomwalk_ts} fits a random walk model using
\code{\link[forecast]{rwf}}

\code{randomwalk_one_step} uses \code{\link[forecast]{rwf}} to fit a
random walk model (with or without drift) and make a single one-step
forecast.
}
