% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting-ranfor.R
\name{ranfor_ts}
\alias{ranfor_ts}
\title{Make forecasts using a random forest model}
\usage{
ranfor_ts(timeseries, num_ahead = 5, level = 95)
}
\arguments{
\item{timeseries}{the time series to forecast}

\item{num_ahead}{the number of points at the end of the time series to
forecast}

\item{level}{Confidence level for prediction intervals.}
}
\value{
a data.frame of the mean forecasts, the observed values, and the
lower and upper CI levels (if an error occurs, then just NA values)
}
\description{
Fit a random forest model with the output as the 1-step ahead
value and the predictors as the 5 previous lags. This is functionally
similar to the time-delay embedding approach with an embedding of 5, but
using \code{\link[randomForest]{randomForest}} to estimate the forecast
function.
}
