% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pe_code.R
\name{rolling_PE}
\alias{rolling_PE}
\title{Entropy of a distribution of counts.}
\usage{
rolling_PE(x, weighted, word_length, tau, tie_method, noise_amount = NA,
  windowsize = 10)
}
\arguments{
\item{wd}{The distribution of counts (e.g., of words).}
}
\value{
The entropy of the distribution.
}
\description{
Entropy of a distribution of counts.
}
\details{
Words containing NAs are ignored.
}
