% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting-sts.R
\name{sts_ts}
\alias{sts_ts}
\title{Structural time series model}
\usage{
sts_ts(timeseries, num_ahead = 5, level = 95, frequency = 1)
}
\arguments{
\item{timeseries}{the time series to forecast}

\item{num_ahead}{the number of points at the end of the time series to
forecast}

\item{level}{Confidence level for prediction intervals.}

\item{frequency}{the number of observations per unit of time.}
}
\value{
a data.frame of the mean forecasts, the observed values, and the
lower and upper CI levels (if an error occurs, then just NA values)
}
\description{
Fit a time series model using \code{\link{StructTS}} and
make forecasts.
}
