% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting-arima.R
\name{arima_fracdiff_ts}
\alias{arima_fracdiff_ts}
\title{Make forward forecasts using a fractionally-differences ARIMA}
\usage{
arima_fracdiff_ts(ts, num_ahead = 5)
}
\arguments{
\item{ts}{a single time series}

\item{num_ahead}{the number of points at the end of the time series to forecast}
}
\value{
a data.frame of the mean forecasts, the observed values, and the
lower and upper 95 CI levels (if an error occurs, then just NA values)
}
\description{
This function performs the forecasting according to the
approach in Ward et al. 2014, where the end of the time series is trimmed,
and a forecasting model is fit to the remainder of the data. This model is
then used to make 1-step ahead forecasts, projecting the forecasts forward
for the end of the time series. (i.e. the first forecast is fed into the
model to make the next forecast, and so forth).)
}
