% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting-arima.R
\name{arima_ts}
\alias{arima_ts}
\alias{randomwalk_ts,}
\alias{arima_fracdiff_ts}
\alias{randomwalk_ts}
\title{Make forecasts using AR / ARIMA}
\usage{
arima_ts(timeseries, num_ahead = 5, order = c(1, 0, 0), level = 95)

arima_fracdiff_ts(timeseries, num_ahead = 5, level = 95)

randomwalk_ts(timeseries, num_ahead = 5, drift = FALSE, level = 95)
}
\arguments{
\item{timeseries}{the time series to forecast}

\item{num_ahead}{the number of points at the end of the time series to
forecast}

\item{order}{A specification of the non-seasonal part of the ARIMA
    model: the three integer components \eqn{(p, d, q)} are the AR order, the
    degree of differencing, and the MA order.}

\item{level}{Confidence level for prediction intervals.}

\item{drift}{Logical flag. If TRUE, fits a random walk with drift model.}
}
\value{
a data.frame of the mean forecasts, the observed values, and the
lower and upper CI levels (if an error occurs, then just NA values)
}
\description{
\code{arima_ts} fits an arima model

\code{arima_fracdiff_ts} fits a step-wise fractionally differenced
auto-arima

\code{randomwalk_ts} fits a random walk model
}
