% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting-ets.R
\name{ets_ts}
\alias{ets_ts}
\alias{sts_ts}
\title{Make forecasts using an exponentially smoothed or structural time
series mdoel}
\usage{
ets_ts(timeseries, num_ahead = 5, level = 95, frequency = 1)

sts_ts(timeseries, num_ahead = 5, level = 95, frequency = 1)
}
\arguments{
\item{timeseries}{the time series to forecast}

\item{num_ahead}{the number of points at the end of the time series to
forecast}

\item{level}{Confidence level for prediction intervals.}

\item{frequency}{the number of observations per unit of time.}
}
\value{
a data.frame of the mean forecasts, the observed values, and the
lower and upper CI levels (if an error occurs, then just NA values)
}
\description{
Fit a time series model and make forecasts. The frequency of the
data is set a priori, as opposed to estimating the parameter from the data.

\code{ets_ts} fits an exponentially smoothed time series model using
\code{\link[forecast]{ets}}

\code{sts_ts} fits a structural model for a time series by
maximum likelihood using \code{\link{StructTS}}
}
