% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting-ets.R
\name{ets_ts}
\alias{ets_ts}
\title{Make forecasts using exponentially smoothed time series}
\usage{
ets_ts(ts, num_ahead = 5, level = 95, frequency = 1)
}
\arguments{
\item{ts}{the time series to forecast}

\item{num_ahead}{the number of points at the end of the time series to
forecast}

\item{level}{the CI level to include}

\item{frequency}{the number of observations per unit of time.}
}
\value{
a data.frame of the mean forecasts, the observed values, and the
lower and upper CI levels (if an error occurs, then just NA values)
}
\description{
Fit an exponentially smoothed time series and make forecasts.
Here, the frequency of the data is set a priori, as opposed to estimating
the parameter from the data
}
