% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-expectations.R
\name{expect_forecasts}
\alias{expect_forecasts}
\title{Check if object is in valid forecasts format}
\usage{
expect_forecasts(output, columns = c("observed", "predicted"),
  known_hash = "cca7c70d85")
}
\arguments{
\item{output}{the object of forecasts to be checked}

\item{columns}{the expecte columns for \code{output} (defaults to
\code{columns = c("observed", "predicted")})}

\item{known_hash}{a known_hash for \code{output}}
}
\description{
Check if object is in valid forecasts format
}
