% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting-utils.R
\name{forecast_wrapper}
\alias{forecast_wrapper}
\title{A wrapper function for forecasting methods}
\usage{
forecast_wrapper(fun, ts, num_ahead = 5, ...)
}
\arguments{
\item{fun}{a}

\item{ts}{the time series to forecast}

\item{num_ahead}{the number of points at the end of the time series to
forecast}

\item{...}{arguments to pass to \code{fun}}
}
\value{
a data.frame of the observed and predicted values, with any other
remaining values returned by \code{fun}. If and error occurs, then NA values for
the observed and predicted
}
\description{
This function faciliates the forecasting approach in Ward et al.
2014:
\enumerate{
\item fit the model to all but the last \code{num_ahead} time points
\item make 1-step ahead forecasts, and feed each forecast in as the
observed, until \code{num_ahead} total forecasts are made
}
}
