% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lpi_data.R
\name{get_LPI_data}
\alias{get_LPI_data}
\title{Read in the LPI time series}
\usage{
get_LPI_data(data_file = system.file("extdata", "example_data.zip",
  package = "rlpi"), min_time_series_length = 25)
}
\value{
list of two dataframes (one with abundance data, the other with
covariate data) and one list of metadata.
}
\description{
Import LPI abundance time series that are included with the
\code{\link{rlpi}} package
}
