% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasting-nnet.R
\name{nnet_ts}
\alias{nnet_ts}
\title{Make forecasts using neural network time series model}
\usage{
nnet_ts(timeseries, num_ahead = 5, level = 95, m = 1, size = 1,
  ...)
}
\arguments{
\item{timeseries}{the time series to forecast}

\item{num_ahead}{the number of points at the end of the time series to
forecast}

\item{level}{Confidence level for prediction intervals.}

\item{m}{ embedding dimension, time delay, forecasting steps }

\item{size}{number of hidden units in the neural network}

\item{...}{arguments to be passed to \code{\link[tsDyn]{predict.nlar}}}
}
\value{
a data.frame of the mean forecasts, the observed values, and the
lower and upper CI levels (if an error occurs, then just NA values)
}
\description{
\code{nnet_ts} fits a neural network time series model with the
specified parameters
}
