% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pe_code.R
\name{PE}
\alias{PE}
\title{Calculate the permuation entropy of a times series}
\usage{
PE(x, weighted, word_length, tau, tie_method, noise_amount = NA)
}
\arguments{
\item{wd}{The distribution of counts (e.g., of words).}
}
\value{
The entropy of the distribution.
}
\description{
Calculate the permuation entropy of a times series
}
\details{
Words containing NAs are ignored.
}
\examples{
x <- rnorm(1000)
wd <- word_distribution(x, 3, 1, order)
PE(wd)
}
