% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pe_code.R
\name{word_distribution}
\alias{word_distribution}
\title{Get the unweighted distribution of "word"s in a time series.}
\usage{
word_distribution(x_emb, tie_method)
}
\arguments{
\item{tie_method}{The method for dealing with tied values; these are the methods used by the rank()
function of the base package. Use "average" to treat ties as the same value,
"first" to treat ties as different with first value being treated as larger than second,
"last" to treat ties as different with last value being treated as larger than second,
"random" to give ties random rank.
@return The word distribution.}

\item{x}{A time series of real numbers.}

\item{word_length}{The word length, also known as the permutation order.}

\item{tau}{Time lag (not implemented).}
}
\description{
Get the unweighted distribution of "word"s in a time series.
}
\examples{
\dontrun{
x <- rnorm(1000)
word_distribution(x, 3, 1, order)
word_distribution <- function(x_emb, tie_method) {
  words <- apply(x_emb, 1, function(x) paste(rev(rank(x, ties.method=tie_method)), collapse="-"))
  table(words)/nrow(x_emb)
}
}
}
