% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pe_code.R
\name{word_distribution}
\alias{word_distribution}
\alias{weighted_word_distribution}
\title{Compute the distribution of "word"s in a time series.}
\usage{
word_distribution(x_emb, ties.method = "first", weighted = TRUE)

weighted_word_distribution(x_emb, ties.method)
}
\arguments{
\item{x_emb}{A time series of real numbers.}

\item{ties.method}{a character string specifying how ties are treated,
    see \sQuote{Details}; can be abbreviated.}

\item{weighted}{whether to weight by the variance in each word}
}
\value{
a table of the word distribution
}
\description{
\code{weighted_word_distribution} is a wrapper that sets \code{weighted}
to \code{TRUE}
}
\details{
#' Weighted distribution of "word"s in a time series.
#'
#' @param x A time series of real numbers.
#' @param word_length The word length, also known as the permutation order.
#' @param tau Time lag (not implemented).
#' @param tie_method The method for dealing with tied values; these are the methods used by the rank()
#' function of the base package. Use "average" to treat ties as the same value,
#' "first" to treat ties as different with first value being treated as larger than second,
#' "last" to treat ties as different with last value being treated as larger than second,
#' "random" to give ties random rank,
#' "noise" to add some noise to the time series to break ties.
#' @param noise_amount How much noise to add to the time series; only used for method "noise".
#' Random numbers from a uniform distribution with maximum 1 * 10^(-noise_amount-1) and minimum zero.
#'  @return The word distribution.
}
